pyproject.toml
[tool.poetry]
name = "BotAInik"
version = "0.1.0"
description = "RAG to prepare for DS interviews, the project was carried out within the framework of the ITMO discipline Introduction to LLM"
authors = ["Nikita Gorbunov <nikita.gorbunov.dev@yandex.com>", "Anna Sidorova <anna.pav.sidorova@gmail.com>", "GlebIsrailevich <glebusai308@gmail.com>"]
readme = "README.md"
packages = [
  { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.318"
openai = "^0.28.0"
requests = "^2.31.0"
PyYAML = "^6.0"
chromadb = "^0.4.10"
pandas = "^2.1.0"
fastapi = "^0.115.6"
sentence-transformers = "^2.2.2"
pypdf = "^3.15.2"
python-docx = "^0.8.11"
openpyxl="^3.1.5"
flask = "^3.1.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.5"
black = "^23.9.1"
pre-commit = "^4.0.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]

[tool.black]
line-length = 99
target-version = ["py310"]